Update bigquery_examples.tf
## Summary

This PR adds the `roles/bigquery.connectionUser` IAM permission to the `explore-assistant-bq-sa` service account in the `bigquery_examples.tf` file. This permission is necessary for the service account to properly connect to BigQuery.

## Changes

- Added `roles/bigquery.connectionUser` IAM permission to the `explore-assistant-bq-sa` service account in the `bigquery_examples.tf` file.

## Reason for Change

While using this resource locally, I encountered an error related to BigQuery connection permissions. Adding the `roles/bigquery.connectionUser` permission resolved the issue. This change ensures that the service account has the necessary permissions to connect to BigQuery, preventing similar issues for other users.

## Testing

Tested locally by applying the Terraform configuration and verifying that the service account can successfully connect to BigQuery without errors.

## Related Issues

N/A